[Wooracle] update interface and param type.

+++ miscellaneous/dex_prod_and_audit/interfaces/IOracle.sol
@@ -43,8 +43,8 @@ interface IOracle {
         view
         returns (
             uint256 latestPrice,
-            uint64 spread,
-            uint64 coefficient,
+            uint256 spread,
+            uint256 coefficient,
             bool feasible
         );
